Simplify modal logics
