+++ cypress/e2e/alert.spec.cy.ts
@@ -27,7 +27,7 @@ beforeEach(() => {
   });
 
   cy.visit("http://localhost:3000/");
-  cy.get("button[id='createButton']").click();
+  cy.get("button[id='newGraphButton']").click();
 
   const intelligentAIButton = cy.get(".modal-content>div>div>button");
   intelligentAIButton.click();
